Removed seperate boundaries file
